requirements.txt
pandas==2.3.0
requests==2.32.4
PyMuPDF==1.26.1
openai==1.88.0
python-dotenv==1.1.0
matplotlib==3.10.3